Move function to web3-utils and more  comments on processVote
